移除过期或无意义脚本
