+++ backUp/jd_esManager.js
