+++ backUp/jd_jxnc.js
