+++ backUp/jd_vivo.js
